# swagger/swagger.yaml
basePath: /
definitions:
...
  dto.UserInsertDto:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      birthdate:
        example: "2006-01-02T15:04:05Z"
        type: string
      city:
        example: São Paulo
        type: string
      document:
        example: "12345678900"
        type: string
      email:
        example: claudio@example.com
        type: string
      name:
        example: Claúdio
        type: string
      pass:
        example: Senhasegur@123
        type: string
      phone:
        example: "21912345678"
        type: string
      role:
        example: developer
        type: string
      state:
        example: São Paulo
        type: string
      type:
        example: fisica
        type: string
    type: object
  dto.UserLoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.UserSSODto:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
  entity.Address:
    properties:
      address:
        type: string
      city:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      state:
        type: string
      userId:
        type: string
    type: object
...
  entity.User:
    properties:
      addresses:
        $ref: '#/definitions/entity.Address'
      avatar_url:
        type: string
      birthdate:
        type: string
      createdAt:
        type: string
      document:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      pass:
        type: string
      phone:
        type: string
      pushNotificationsEnabled:
        type: boolean
      role:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:3000/api
info:
  contact:
    name: DevHatt
    url: https://github.com/devhatt
  description: Esta página se destina a documentação da API do projeto PetDex Backend
  license:
    name: MIT license
    url: https://github.com/devhatt/pet-dex-backend?tab=MIT-1-ov-file#readme
  title: 'PetDex: Documentação API'
  version: "1.0"
paths:
...
  /user/{provider}/login:
    post:
      consumes:
      - application/json
      description: Logs in a user using a specified provider (SSO) and returns a JWT
        token
      parameters:
      - description: The provider for Single Sign-On (e.g., google, facebook)
        in: path
        name: provider
        required: true
        type: string
      - description: User login information with SSO
        in: body
        name: UserSSODto
        required: true
        schema:
          $ref: '#/definitions/dto.UserSSODto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: User login with provider
      tags:
      - User
  /user/{userID}:
    get:
      consumes:
      - application/json
      description: Retrieves User details based on the User ID provided as a parameter.
      parameters:
      - description: ID of the User to be retrieved
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Find User by ID
      tags:
      - User
...
  /user/create-account:
    post:
      consumes:
      - application/json
      description: Creates user and insert into the database
      parameters:
      - description: User object information to create
        in: body
        name: userDto
        required: true
        schema:
          $ref: '#/definitions/dto.UserInsertDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Creates user
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Logs in a user and returns a JWT token
      parameters:
      - description: User login information
        in: body
        name: userLoginDto
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: User login
      tags:
      - User
swagger: "2.0"